requirements.txt
numpy>=1.24,<3
Pillow>=10.0
pytesseract>=0.3.10
scikit-learn>=1.3,<2
streamlit>=1.30,<2
pydantic>=2.7,<3
pymupdf>=1.24,<2
